requirements.txt
mysql-connector-python==8.3.0
anyio==3.7.1
click==8.0.3
colorama==0.4.4
fastapi==0.105.0
greenlet==3.0.3
h11==0.14.0
idna==3.3
pydantic==2.5.2
pydantic_core==2.14.5
PyJWT==2.3.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1